add helper function util, create setPreview helper, moved all validations to validation.js, removed unused code
