+++ backend/routes/api/users.js
@@ -4,37 +4,9 @@ const bcrypt = require("bcryptjs");
 const { setTokenCookie, requireAuth } = require("../../utils/auth");
 const { User } = require("../../db/models");
 
-const { check } = require("express-validator");
-const { handleValidationErrors } = require("../../utils/validation");
+const { validateSignup } = require("../../utils/validation");
 
 const router = express.Router();
-
-const validateSignup = [
-  check("firstName")
-    .exists({ checkFalsy: true })
-    .isLength({ min: 1, max: 16 })
-    .withMessage("Please provide a First Name between 1-16 characters"),
-  check("lastName")
-    .exists({ checkFalsy: true })
-    .isLength({ min: 1, max: 16 })
-    .withMessage("Please provide a Last Name between 1-16 characters"),
-  check("email")
-    .exists({ checkFalsy: true })
-    .isEmail()
-    .withMessage("Please provide a valid email."),
-  check("username")
-    .exists({ checkFalsy: true })
-    .isAlphanumeric()
-    .withMessage("Please provide a valid username")
-    .isLength({ min: 4 })
-    .withMessage("Please provide a username with at least 4 characters."),
-  check("username").not().isEmail().withMessage("Username cannot be an email."),
-  check("password")
-    .exists({ checkFalsy: true })
-    .isLength({ min: 6 })
-    .withMessage("Password must be 6 characters or more."),
-  handleValidationErrors,
-];
 
 /* Sign up */
 router.post("/", validateSignup, async (req, res, next) => {
